chore(remove low quality code): moving low quality code into internal for now
