+++ internal/README.md
@@ -1,0 +1,16 @@
+
+# Highly Experimental code; note meant to be used or referenced.
+
+## record buffer
+
+Store records in any byte stream.
+
+## writer cache
+
+Keeps tracks of multiple writer factories and closes them based on timeouts or LRU.
+
+## concurrent
+
+WIP/Playground code - DO NOT USE.
+
+Concurrent ordered execution of jobs in one or more steps. Similar to concurrent map in functional languages
