+++ internal/iterator/dedup_iterator.go
@@ -1,13 +1,15 @@
 package iterator
 
+import "github.com/kvanticoss/goutils/iterator"
+
 // DeduplicateRecordIterators works like the unix command uniq where if two records are equal (using .Less(other) bool as comparison func)
 // only the first are emitted. If not both records implement .Less(interface{})bool DeduplicateRecordIterators is a NOP (with overhead).
 // Nil records are never considered equal
-func DeduplicateRecordIterators(it RecordIterator) RecordIterator {
+func DeduplicateRecordIterators(it iterator.RecordIterator) iterator.RecordIterator {
 	var previousRecord interface{}
 	return func() (interface{}, error) {
 		rec, err := it()
-		
+
 		for err == nil && areEqual(rec, previousRecord) {
 			previousRecord = rec
 			rec, err = it()
