+++ internal/iterator/lesser_iterator.go
