+++ internal/recordwriter/json_conts.go
@@ -1,0 +1,3 @@
+package recordwriter
+
+var jsonRecordDelimiter = []byte("\n")
