+++ iterator/log_middleware.go
@@ -12,14 +12,3 @@ func LogRecordIterator(it RecordIterator, pattern string) RecordIterator {
 		return r, err
 	}
 }
-
-// LogLesserterator prints the contents of the record prior to returning it
-// using the pattern as the fmt-directive where the first argument is the records
-// second is the error as such. log.Printf(pattern, r, err)
-func LogLesserterator(it LesserIterator, pattern string) LesserIterator {
-	return func() (Lesser, error) {
-		r, err := it()
-		log.Printf(pattern, r, err)
-		return r, err
-	}
-}
